requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appdirs==1.4.3            # via gmusicapi
attrs==19.3.0             # via glom, pytest
beautifulsoup4==4.9.0     # via mechanicalsoup
boltons==20.1.0           # via face, glom
certifi==2020.4.5.1       # via requests
cffi==1.14.0              # via cryptography
chardet==3.0.4            # via requests
click==7.1.2              # via pip-tools
cryptography==2.9.2       # via -r requirements.in
decorator==4.4.2          # via gmusicapi
face==20.1.1              # via glom
glom==20.5.0              # via -r requirements.in
gmusicapi==13.0.0         # via -r requirements.in
gpsoauth==0.4.1           # via gmusicapi
httplib2==0.17.3          # via oauth2client
idna==2.9                 # via requests
importlib-metadata==1.6.0  # via pluggy, pytest
lxml==4.5.0               # via mechanicalsoup
mechanicalsoup==0.12.0    # via gmusicapi
more-itertools==8.2.0     # via pytest
mutagen==1.44.0           # via gmusicapi
oauth2client==4.1.3       # via gmusicapi
packaging==20.3           # via pytest
pip-tools==5.1.2          # via -r requirements.in
pluggy==0.13.1            # via pytest
proboscis==1.2.6.0        # via gmusicapi
prompt-toolkit==3.0.5     # via -r requirements.in
protobuf==3.11.3          # via gmusicapi
py==1.8.1                 # via pytest
pyasn1-modules==0.2.8     # via oauth2client
pyasn1==0.4.8             # via oauth2client, pyasn1-modules, rsa
pycparser==2.20           # via cffi
pycryptodomex==3.9.7      # via gpsoauth
pyjwt==1.7.1              # via -r requirements.in
pyparsing==2.4.7          # via packaging
pytest==5.4.2             # via -r requirements.in
python-dateutil==2.8.1    # via gmusicapi
requests==2.23.0          # via -r requirements.in, gmusicapi, gpsoauth, mechanicalsoup
rsa==4.0                  # via oauth2client
six==1.14.0               # via cryptography, mechanicalsoup, oauth2client, packaging, pip-tools, protobuf, python-dateutil
soupsieve==2.0            # via beautifulsoup4
urllib3==1.25.9           # via requests
validictory==1.1.2        # via gmusicapi
wcwidth==0.1.9            # via prompt-toolkit, pytest
zipp==3.1.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
